chore(Queries): qt editor view to the different file [YTFRONT-4802]
