+++ packages/ui/src/ui/pages/query-tracker/QueryEditor/QueryEditor.scss
@@ -26,49 +26,4 @@
         /* In order to override flex basis from split component */
         flex-basis: initial !important;
     }
-
-    &__query {
-        width: 100%;
-        height: 100%;
-        overflow: visible;
-    }
-
-    &__editor {
-        flex: 1;
-        border-bottom: 1px solid var(--dark-divider);
-        width: 100%;
-        height: calc(100% - 40px);
-        overflow: visible;
-    }
-
-    &__actions {
-        padding: 6px 20px;
-        display: flex;
-        flex-direction: row;
-        background: var(--main-background);
-
-        .query-run-action {
-            width: 100%;
-            display: flex;
-            align-items: center;
-            gap: 0 8px;
-            margin-right: 12px;
-        }
-
-    }
-
-    &__error-line {
-        background: var(--error-background);
-    }
-
-    .selected-text {
-        z-index: 2;
-        position: absolute;
-        pointer-events: none;
-        opacity: 0.8;
-    }
-
-    .view-lines {
-        z-index: 3;
-    }
 }
