+++ packages/ui/src/ui/pages/query-tracker/QueryEditor/QueryEditorView.scss
@@ -1,0 +1,44 @@
+.yt-qt-query-editor-view {
+    width: 100%;
+    height: 100%;
+    overflow: visible;
+
+    &__editor {
+        flex: 1;
+        border-bottom: 1px solid var(--dark-divider);
+        width: 100%;
+        height: calc(100% - 40px);
+        overflow: visible;
+    }
+
+    &__error-line {
+        background: var(--error-background);
+    }
+
+    &__actions {
+        padding: 6px 20px;
+        display: flex;
+        flex-direction: row;
+        background: var(--main-background);
+
+        .query-run-action {
+            width: 100%;
+            display: flex;
+            align-items: center;
+            gap: 0 8px;
+            margin-right: 12px;
+        }
+
+    }
+
+    .selected-text {
+        z-index: 2;
+        position: absolute;
+        pointer-events: none;
+        opacity: 0.8;
+    }
+
+    .view-lines {
+        z-index: 3;
+    }
+}
